+++ R/docx_img.R
@@ -16,6 +16,11 @@ docx_reference_img <- function( x, src){
   x <- part_reference_img(x, src, "doc_obj")
   x
 }
+docx_reference_hyperlink <- function( x, href){
+  for(hr in href)
+    x <- part_reference_hyperlink(x, hr, "doc_obj")
+  x
+}
 
 part_reference_img <- function( x, src, part){
   src <- unique( src )
@@ -24,6 +29,16 @@ part_reference_img <- function( x, src, part){
   img_path <- file.path(x$package_dir, "word", "media")
   dir.create(img_path, recursive = TRUE, showWarnings = FALSE)
   file.copy(from = src, to = file.path(x$package_dir, "word", "media", basename(src)))
+
+  x
+}
+part_reference_hyperlink <- function( x, href, part){
+  hyperlink_id <- paste0("rId", x[[part]]$relationship()$get_next_id())
+  x[[part]]$relationship()$add(
+    id = hyperlink_id,
+    type = "http://schemas.openxmlformats.org/officeDocument/2006/relationships/hyperlink",
+    target = htmlEscapeCopy(href),
+    target_mode = "External" )
 
   x
 }
@@ -59,6 +74,27 @@ wml_part_link_images <- function(x, str, part){
     src_nodes <- xml_find_all(doc, xpth)
     blip_nodes <- xml_find_all(src_nodes, "wp:inline/a:graphic/a:graphicData/pic:pic/pic:blipFill/a:blip")
     xml_attr(blip_nodes, "r:embed") <- ref$id[id]
+  }
+  as.character(doc)
+}
+
+wml_link_hyperlink <- function(x, str){
+  wml_part_link_hyperlink(x, str, "doc_obj")
+}
+
+wml_part_link_hyperlink <- function(x, str, part){
+  ref <- x[[part]]$relationship()$get_data()
+
+  ref <- ref[ref$target_mode %in% "External",]
+
+  doc <- as_xml_document(str)
+  for(id in seq_along(ref$ext_src) ){
+    xpth <- paste0("//w:hyperlink",
+                   sprintf( "[contains(@r:id,'%s')]", ref$target[id])
+                   )
+
+    src_nodes <- xml_find_all(doc, xpth)
+    xml_attr(src_nodes, "r:id") <- ref$id[id]
   }
   as.character(doc)
 }